Simplify _get_cert_url
Just tweaked this to remove deep nesting and replace it with early abortions.
All functionality is maintained, but this does highlight that there's an
unaccounted for code path after the last `if`. If the parts don't line up, we
just return None. Should we instead log a warning or...something?