Update the notification type key
Incoming payload from AWS SNS contains a message, which is a
JSON string.
This message used to have as a key "eventType" to designate if the event
was a Delivery or a Bounce for example.

The key no longer exists and is now "notificationType".

This leads every notification sent to fall in the "unknown event type"
case in SNSEventWebhookView.

Closes #247