Handle potential notificationType in bounce and complaint events
According to the official docs[^1] the top level key is either `eventType` or
`notificationType`, depending on whether or not users have configured event
publishing.

The current docs for django-ses recommend setting up a configuration set and
event publishing, but as mentioned in #248, it's possible to use django-ses
with the less-complete notification system.

This PR changes django-ses to handle either `notificationType` or `eventType`,
preferring the current `eventType` if it exists.

[^1]: https://docs.aws.amazon.com/ses/latest/dg/notification-contents.html#top-level-json-object